feat: refactor recommendation modal and situation create page, implement CreateSituationForm component
